+++ src/dreamstate/core/signals/useSignals.ts
@@ -1,15 +1,16 @@
 import { useEffect } from "react";
 
-import { EMPTY_ARR } from "@/dreamstate/core/internals";
 import { subscribeToSignals } from "@/dreamstate/core/signals/subscribeToSignals";
 import { unsubscribeFromSignals } from "@/dreamstate/core/signals/unsubscribeFromSignals";
 import { TCallable, TSignalListener, TSignalType } from "@/dreamstate/types";
 
 /**
  * Hook for signals listening and custom UI handling.
+ * Provided dependencies will trigger function re-subscription.
  */
 export function useSignals<T extends TSignalType = TSignalType, D = undefined>(
-  subscriber: TSignalListener<T, D>
+  subscriber: TSignalListener<T, D>,
+  dependencies: Array<any>
 ): void {
   useEffect(function(): TCallable {
     subscribeToSignals(subscriber);
@@ -17,5 +18,5 @@ export function useSignals<T extends TSignalType = TSignalType, D = undefined>(
     return function(): void {
       unsubscribeFromSignals(subscriber);
     };
-  }, EMPTY_ARR);
+  }, dependencies);
 }